+++ test/notification.test.js
@@ -9,53 +9,36 @@ const apiBootstrap = require('../lib/apiBootstrap');
 require('./src/userSchema');
 require('./src/inviteSchema');
 require('./src/notificationSchema');
+require('./src/messageSchema');
+
 const notification = require('../lib/notification');
 
 
 const connectionData = require('./src/connectionData');
-//
-// // Test data
-// let emailToInvite = 'tiago@tiagogouvea.com.br';
-// let phoneToInvite = '32988735683';
-// let emailToInviteAndSignup = "tiago@appmasters.io";
-// let personsToInvite = [
-//     {email: 'email@email.com', phone: '99837495'},
-//     {email: emailToInviteAndSignup, phone: '99833743'}
-// ];
-//
-// let invite = null;
-// let invites = null;
-// let user = null;
-// let newUser = null;
-// let newUsers = null;
 
 jasmine.DEFAULT_TIMEOUT_INTERVAL = 10000;
 
-// test('ApiBootstrap node-lib', () => {
-//     let result = apiBootstrap.setup(app, envs, packag, passport);
-//     apiBootstrap.listen(app);
-//     expect(result).toBe(true);
-// });
-//
-// test('Create connection and one user', async () => {
-//     let db = connectionData.createConnectionAndSchemas();
-//     await db.connection.model('user').find({}).remove();
-//     user = await connectionData.createUserRecords(1);
-//     expect(user).not.toBeNull();
-// });
+test('ApiBootstrap node-lib', () => {
+    let result = apiBootstrap.setup(app, envs, packag, passport);
+    apiBootstrap.listen(app);
+    expect(result).toBe(true);
+});
+
+test('Create connection', async () => {
+    let db = connectionData.createConnectionAndSchemas();
+    expect(db).not.toBeNull();
+});
 
 test('Send notification', async () => {
-    const credential = require('./src/serviceAccountKey');
-    let config = {credential, databaseURL: 'https://good-burger.firebaseio.com'};
-    let result = notification.setup(config);
-    expect(result).toBe(true);
-    const users = {
-        token: 'flHQT5RETSg:APA91bG3Go3LXcuV2V-ATLJFDLzeQ5z6mQA22JEuyRedWio1Jfj6VAXV7qbgoofu6TKZPWB86Li74PQT26Drth6pgwKKhtxr5-VQNui-iusvstcVHT7mtMv0arquyRWpNhAFWJAXUY_h'
-    };
+    const users =
+        {
+            _id: '5980d6e14a54b900113ff905',
+            token: 'flHQT5RETSg:APA91bG3Go3LXcuV2V-ATLJFDLzeQ5z6mQA22JEuyRedWio1Jfj6VAXV7qbgoofu6TKZPWB86Li74PQT26Drth6pgwKKhtxr5-VQNui-iusvstcVHT7mtMv0arquyRWpNhAFWJAXUY_h'
+        };
     let payload = {
         notification: {
-            title: 'Teste',
-            body: 'Teste de notificação'
+            title: 'Olá',
+            body: 'Bem vindo!'
         },
         data: {
             content: 'Only strings goes here',
@@ -65,5 +48,4 @@ test('Send notification', async () => {
     notification.send(users, payload);
     // expect(result.successCount).toBe(1);
 });
-